Separated infix rewriter from writers.cc, updated exists/can_replace
